feat: add study list component
